+++ cargodash/__init__.py
@@ -17,4 +17,4 @@ __all__ = [
     "LocalHFChatClient", "LocalVLLMChatClient", "LLMCall",
 ]
 
-__version__ = "1.0.5"
+__version__ = "1.0.6"